+++ graphql_server/flask/graphqlview.py
@@ -39,6 +39,7 @@ class GraphQLView(View):
     validation_rules = None
     execution_context_class = None
     batch = False
+    jinja_env = None
     subscriptions = None
     headers = None
     default_query = None
@@ -131,7 +132,7 @@ class GraphQLView(View):
                     graphiql_version=self.graphiql_version,
                     graphiql_template=self.graphiql_template,
                     graphiql_html_title=self.graphiql_html_title,
-                    jinja_env=None,
+                    jinja_env=self.jinja_env,
                 )
                 graphiql_options = GraphiQLOptions(
                     default_query=self.default_query,
