+++ tests/flask/test_graphiqlview.py
@@ -1,25 +1,14 @@
 import pytest
 from flask import url_for
+from jinja2 import Environment
 
 from .app import create_app
 
 
-@pytest.fixture
-def app():
-    # import app factory pattern
-    app = create_app(graphiql=True)
-
-    # pushes an application context manually
-    ctx = app.app_context()
-    ctx.push()
-    return app
-
-
-@pytest.fixture
-def client(app):
-    return app.test_client()
-
-
+@pytest.mark.parametrize(
+    "app",
+    [create_app(graphiql=True), create_app(graphiql=True, jinja_env=Environment())],
+)
 def test_graphiql_is_enabled(app, client):
     with app.test_request_context():
         response = client.get(
@@ -28,6 +17,10 @@ def test_graphiql_is_enabled(app, client):
     assert response.status_code == 200
 
 
+@pytest.mark.parametrize(
+    "app",
+    [create_app(graphiql=True), create_app(graphiql=True, jinja_env=Environment())],
+)
 def test_graphiql_renders_pretty(app, client):
     with app.test_request_context():
         response = client.get(
@@ -45,6 +38,10 @@ def test_graphiql_renders_pretty(app, client):
     assert pretty_response in response.data.decode("utf-8")
 
 
+@pytest.mark.parametrize(
+    "app",
+    [create_app(graphiql=True), create_app(graphiql=True, jinja_env=Environment())],
+)
 def test_graphiql_default_title(app, client):
     with app.test_request_context():
         response = client.get(url_for("graphql"), headers={"Accept": "text/html"})
@@ -52,7 +49,13 @@ def test_graphiql_default_title(app, client):
 
 
 @pytest.mark.parametrize(
-    "app", [create_app(graphiql=True, graphiql_html_title="Awesome")]
+    "app",
+    [
+        create_app(graphiql=True, graphiql_html_title="Awesome"),
+        create_app(
+            graphiql=True, graphiql_html_title="Awesome", jinja_env=Environment()
+        ),
+    ],
 )
 def test_graphiql_custom_title(app, client):
     with app.test_request_context():
