+++ tests/webob/conftest.py
@@ -1,0 +1,13 @@
+import pytest
+
+from .app import Client
+
+
+@pytest.fixture
+def settings():
+    return {}
+
+
+@pytest.fixture
+def client(settings):
+    return Client(settings=settings)
