+++ tests/webob/test_graphiqlview.py
@@ -1,43 +1,36 @@
 import pytest
+from jinja2 import Environment
 
-from .app import Client, url_string
-
-
-@pytest.fixture
-def settings():
-    return {}
+from .app import url_string
 
 
-@pytest.fixture
-def client(settings):
-    return Client(settings=settings)
+@pytest.mark.parametrize(
+    "settings", [dict(graphiql=True), dict(graphiql=True, jinja_env=Environment())]
+)
+def test_graphiql_is_enabled(client):
+    response = client.get(url_string(query="{test}"), headers={"Accept": "text/html"})
+    assert response.status_code == 200
 
 
-@pytest.fixture
-def pretty_response():
-    return (
+@pytest.mark.parametrize(
+    "settings", [dict(graphiql=True), dict(graphiql=True, jinja_env=Environment())]
+)
+def test_graphiql_simple_renderer(client):
+    response = client.get(url_string(query="{test}"), headers={"Accept": "text/html"})
+    assert response.status_code == 200
+    pretty_response = (
         "{\n"
         '  "data": {\n'
         '    "test": "Hello World"\n'
         "  }\n"
         "}".replace('"', '\\"').replace("\n", "\\n")
     )
-
-
-@pytest.mark.parametrize("settings", [dict(graphiql=True)])
-def test_graphiql_is_enabled(client, settings):
-    response = client.get(url_string(query="{test}"), headers={"Accept": "text/html"})
-    assert response.status_code == 200
-
-
-@pytest.mark.parametrize("settings", [dict(graphiql=True)])
-def test_graphiql_simple_renderer(client, settings, pretty_response):
-    response = client.get(url_string(query="{test}"), headers={"Accept": "text/html"})
-    assert response.status_code == 200
     assert pretty_response in response.body.decode("utf-8")
 
 
-@pytest.mark.parametrize("settings", [dict(graphiql=True)])
-def test_graphiql_html_is_not_accepted(client, settings):
+@pytest.mark.parametrize(
+    "settings", [dict(graphiql=True), dict(graphiql=True, jinja_env=Environment())]
+)
+def test_graphiql_html_is_not_accepted(client):
     response = client.get(url_string(), headers={"Accept": "application/json"})
     assert response.status_code == 400
